+++ torchci/test/flakyBotTests/flakyBotUtils.test.ts
@@ -1,478 +1,19 @@
-import dayjs from "dayjs";
-import { parseBody } from "lib/bot/verifyDisableTestIssueBot";
+import * as singleDisableIssue from "lib/flakyBot/singleDisableIssue";
+import * as flakyBotUtils from "lib/flakyBot/utils";
 import { IssueData } from "lib/types";
 import nock from "nock";
-import * as disableFlakyTestBot from "../pages/api/flaky-tests/disable";
-import { deepCopy, handleScope } from "./common";
-import * as utils from "./utils";
-
-nock.disableNetConnect();
-
-const flakyTestA = {
-  file: "file_a.py",
-  invoking_file: "file_a",
-  suite: "suite_a",
-  name: "test_a",
-  numGreen: 4,
-  numRed: 2,
-  workflowIds: ["12345678", "13456789", "14253647"],
-  workflowNames: ["trunk", "periodic", "periodic"],
-  jobIds: [55443322, 55667788, 56789876],
-  jobNames: [
-    "win-cpu-vs-2019 / test",
-    "win-cuda11.3-vs-2019 / test",
-    "win-cuda11.3-vs-2019 / test",
-  ],
-  branches: ["master", "master", "master"],
-};
-
-const flakyTestB = {
-  file: "file_b.py",
-  invoking_file: "file_b",
-  suite: "suite_b",
-  name: "test_b",
-  numGreen: 4,
-  numRed: 2,
-  workflowIds: ["12345678", "13456789", "14253647"],
-  workflowNames: [
-    "win-cpu-vs-2019",
-    "periodic-win-cuda11.3-vs-2019",
-    "periodic-win-cuda11.3-vs-2019",
-  ],
-  jobIds: [55443322, 55667788, 56789876],
-  jobNames: [
-    "win-cpu-vs-2019 / test",
-    "win-cuda11.3-vs-2019 / test",
-    "win-cuda11.3-vs-2019 / test",
-  ],
-  branches: ["main", "main", "main"],
-};
-
-const flakyTestE = {
-  file: "file_e.py",
-  invoking_file: "file_e",
-  suite: "suite_e",
-  name: "test_e",
-  numGreen: 4,
-  numRed: 2,
-  workflowIds: ["12345678", "13456789", "14253647", "15949539"],
-  workflowNames: ["pull", "periodic", "trunk", "pull"],
-  jobIds: [55443322, 55667788, 56789876, 56677889],
-  jobNames: [
-    "win-cpu-vs-2019 / test",
-    "linux-xenial-cuda11.5-py3 / test",
-    "macos-11-x86 / test",
-    "win-cpu-vs-2019 / test",
-  ],
-  branches: ["pr-fix", "master", "master", "another-pr-fx"],
-};
-
-const flakyTestAcrossJobA = {
-  name: "test_conv1d_vs_scipy_mode_same_cuda_complex64",
-  suite: "TestConvolutionNNDeviceTypeCUDA",
-  file: "nn/test_convolution.py",
-  invoking_file: "nn.test_convolution",
-  jobNames: [
-    "linux-focal-rocm5.2-py3.8 / test (default, 1, 2, linux.rocm.gpu)",
-    "linux-focal-rocm5.2-py3.8 / test (default, 1, 2, linux.rocm.gpu)",
-  ],
-  jobIds: [9489898216, 9486287115],
-  workflowIds: ["3466924095", "3465587581"],
-  workflowNames: ["trunk", "trunk"],
-  runAttempts: [1, 1],
-  eventTimes: ["2022-11-15T02:52:20.311000Z", "2022-11-14T22:30:34.492000Z"],
-  branches: ["master", "master"],
-};
-
-const nonFlakyTestA = {
-  name: "test_a",
-  classname: "suite_a",
-  filename: "file_a.py",
-  flaky: false,
-  num_green: 50,
-  num_red: 0,
-};
-
-const nonFlakyTestZ = {
-  name: "test_z",
-  classname: "suite_z",
-  filename: "file_z.py",
-  flaky: false,
-  num_green: 50,
-  num_red: 0,
-};
-
-function mockGetRawTestFile(file: string, content: string) {
-  return nock("https://raw.githubusercontent.com")
-    .get(`/pytorch/pytorch/main/test/${file}`)
-    .reply(200, Buffer.from(content));
-}
-
-describe("Disable Flaky Test Bot Across Jobs", () => {
-  const octokit = utils.testOctokit();
-
-  beforeEach(() => {});
-
-  afterEach(() => {
-    jest.restoreAllMocks();
-  });
-
-  test("Create new issue", async () => {
-    const scope = mockGetRawTestFile(
-      flakyTestAcrossJobA.file,
-      `# Owner(s): ["module: fft"]\nimport blah;\nrest of file`
-    );
-    const scope2 = utils.mockCreateIssue(
-      "pytorch/pytorch",
-      "DISABLED test_conv1d_vs_scipy_mode_same_cuda_complex64 (__main__.TestConvolutionNNDeviceTypeCUDA)",
-      ["Platforms: "],
-      [
-        "skipped",
-        "module: flaky-tests",
-        "module: fft",
-        "module: rocm",
-        "triaged",
-      ]
-    );
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTestAcrossJobA, [], octokit);
-
-    handleScope(scope);
-    handleScope(scope2);
-  });
-
-  test("flaky test associated with an open issue should comment if recent", async () => {
-    let flakyTest = { ...flakyTestAcrossJobA };
-    flakyTest.eventTimes = [dayjs().subtract(1, "hour").toString()];
-    const scope = nock("https://api.github.com")
-      .post("/repos/pytorch/pytorch/issues/1/comments", (body) => {
-        const comment = JSON.stringify(body.body);
-        expect(comment).toContain(
-          "Another case of trunk flakiness has been found"
-        );
-        expect(comment).toContain("appears to contain");
-        expect(comment).toContain("Either the change didn't propogate");
-        return true;
-      })
-      .reply(200, {});
-
-    const issues = [
-      {
-        number: 1,
-        title:
-          "DISABLED test_conv1d_vs_scipy_mode_same_cuda_complex64 (__main__.TestConvolutionNNDeviceTypeCUDA)",
-        html_url: "https://api.github.com/repos/pytorch/pytorch/issues/1",
-        state: "open" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs().toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTest, issues, octokit);
-
-    handleScope(scope);
-  });
-
-  test("flaky test associated with an open issue should NOT comment if NOT recent", async () => {
-    let flakyTest = { ...flakyTestAcrossJobA };
-    flakyTest.eventTimes = [dayjs().subtract(5, "hour").toString()];
-
-    const issues = [
-      {
-        number: 1,
-        title:
-          "DISABLED test_conv1d_vs_scipy_mode_same_cuda_complex64 (__main__.TestConvolutionNNDeviceTypeCUDA)",
-        html_url: "https://api.github.com/repos/pytorch/pytorch/issues/1",
-        state: "open" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs().toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTest, issues, octokit);
-  });
-
-  test("flaky test associated with a closed issue should reopen issue and comment if recent", async () => {
-    let flakyTest = { ...flakyTestAcrossJobA };
-    flakyTest.eventTimes = [dayjs().subtract(1, "hour").toString()];
-    const scope = nock("https://api.github.com")
-      .patch("/repos/pytorch/pytorch/issues/1", (body) => {
-        expect(body).toMatchObject({ state: "open" });
-        return true;
-      })
-      .reply(200, {})
-      .post("/repos/pytorch/pytorch/issues/1/comments", (body) => {
-        const comment = JSON.stringify(body.body);
-        expect(comment).toContain(
-          "Another case of trunk flakiness has been found"
-        );
-        expect(comment).toContain("Reopening");
-        return true;
-      })
-      .reply(200, {});
-
-    const issues = [
-      {
-        number: 1,
-        title:
-          "DISABLED test_conv1d_vs_scipy_mode_same_cuda_complex64 (__main__.TestConvolutionNNDeviceTypeCUDA)",
-        html_url: "https://api.github.com/repos/pytorch/pytorch/issues/1",
-        state: "closed" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs().toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTest, issues, octokit);
-
-    handleScope(scope);
-  });
-
-  test("flaky test associated with a closed issue should NOT reopen issue and comment if NOT recent", async () => {
-    let flakyTest = { ...flakyTestAcrossJobA };
-    flakyTest.eventTimes = [dayjs().subtract(5, "hour").toString()];
-
-    const issues = [
-      {
-        number: 1,
-        title:
-          "DISABLED test_conv1d_vs_scipy_mode_same_cuda_complex64 (__main__.TestConvolutionNNDeviceTypeCUDA)",
-        html_url: "https://api.github.com/repos/pytorch/pytorch/issues/1",
-        state: "closed" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs().toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTest, issues, octokit);
-  });
-});
-
-describe("Disable Flaky Test Bot Integration Tests", () => {
-  const octokit = utils.testOctokit();
-
-  beforeEach(() => {});
-
-  afterEach(() => {
-    jest.restoreAllMocks();
-  });
-
-  test("previously undetected flaky test should create an issue", async () => {
-    const scope = mockGetRawTestFile(
-      flakyTestA.file,
-      `# Owner(s): ["module: fft"]\nimport blah;\nrest of file`
-    );
-    const scope2 = utils.mockCreateIssue(
-      "pytorch/pytorch",
-      "DISABLED test_a (__main__.suite_a)",
-      ["Platforms: "],
-      [
-        "skipped",
-        "module: flaky-tests",
-        "module: fft",
-        "module: windows",
-        "triaged",
-      ]
-    );
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTestA, [], octokit);
-
-    handleScope(scope);
-    handleScope(scope2);
-  });
-
-  test("previously undetected flaky test should create an issue on main", async () => {
-    const scope = mockGetRawTestFile(
-      flakyTestB.file,
-      `# Owner(s): ["module: fft"]\nimport blah;\nrest of file`
-    );
-    const scope2 = utils.mockCreateIssue(
-      "pytorch/pytorch",
-      "DISABLED test_b (__main__.suite_b)",
-      ["Platforms:"],
-      [
-        "skipped",
-        "module: flaky-tests",
-        "module: fft",
-        "module: windows",
-        "triaged",
-      ]
-    );
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTestB, [], octokit);
-
-    handleScope(scope);
-    handleScope(scope2);
-  });
-
-  test("flaky test associated with an open issue should comment", async () => {
-    const scope = nock("https://api.github.com")
-      .post("/repos/pytorch/pytorch/issues/1/comments", (body) => {
-        const comment = JSON.stringify(body.body);
-        expect(comment).toContain(
-          "Another case of trunk flakiness has been found"
-        );
-        expect(comment).toContain("Either the change didn't propogate");
-        return true;
-      })
-      .reply(200, {});
-
-    const issues: IssueData[] = [
-      {
-        number: 1,
-        title: "DISABLED test_a (__main__.suite_a)",
-        html_url: "https://api.github.com/repos/pytorch/pytorch/issues/1",
-        state: "open" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs().toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTestA, issues, octokit);
-
-    handleScope(scope);
-  });
-
-  test("flaky test associated with a closed issue should reopen issue and comment", async () => {
-    const scope = nock("https://api.github.com")
-      .patch("/repos/pytorch/pytorch/issues/1", (body) => {
-        expect(body).toMatchObject({ state: "open" });
-        return true;
-      })
-      .reply(200, {})
-      .post("/repos/pytorch/pytorch/issues/1/comments", (body) => {
-        const comment = JSON.stringify(body.body);
-        expect(comment).toContain(
-          "Another case of trunk flakiness has been found"
-        );
-        expect(comment).toContain("Reopening");
-        return true;
-      })
-      .reply(200, {});
-
-    const issues: IssueData[] = [
-      {
-        number: 1,
-        title: "DISABLED test_a (__main__.suite_a)",
-        html_url: "https://api.github.com/pytorch/pytorch/issues/1",
-        state: "closed" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs().toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTestA, issues, octokit);
-
-    handleScope(scope);
-  });
-
-  test("comment and close non flaky test", async () => {
-    const scope = nock("https://api.github.com")
-      .post("/repos/pytorch/pytorch/issues/1/comments", (body) => {
-        const comment = JSON.stringify(body.body);
-        expect(comment).toContain(
-          "Another case of trunk flakiness has been found"
-        );
-        expect(comment).toContain("Either the change didn't propogate");
-        return true;
-      })
-      .reply(200, {})
-      .post("/repos/pytorch/pytorch/issues/1/comments", (body) => {
-        const comment = JSON.stringify(body.body);
-        expect(comment).toContain(
-          "Resolving the issue because the test is not flaky anymore"
-        );
-        return true;
-      })
-      .reply(200, {})
-      .patch("/repos/pytorch/pytorch/issues/1", (body) => {
-        expect(body).toMatchObject({ state: "closed" });
-        return true;
-      })
-      .reply(200, {});
-
-    const issues: IssueData[] = [
-      {
-        number: 1,
-        title: "DISABLED test_a (__main__.suite_a)",
-        html_url: "https://api.github.com/repos/pytorch/pytorch/issues/1",
-        state: "open" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs()
-          .subtract(
-            disableFlakyTestBot.NUM_HOURS_NOT_UPDATED_BEFORE_CLOSING + 1,
-            "hour"
-          )
-          .toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTestA, issues, octokit);
-    // Close the disabled issue if the test is not flaky anymore
-    await disableFlakyTestBot.handleNonFlakyTest(
-      nonFlakyTestA,
-      issues,
-      octokit
-    );
-
-    handleScope(scope);
-  });
-
-  test("do not close non flaky test if it's manual updated recently", async () => {
-    const scope = nock("https://api.github.com")
-      .post("/repos/pytorch/pytorch/issues/1/comments", (body) => {
-        const comment = JSON.stringify(body.body);
-        expect(comment).toContain(
-          "Another case of trunk flakiness has been found"
-        );
-        expect(comment).toContain("Either the change didn't propogate");
-        return true;
-      })
-      .reply(200, {});
-
-    const issues: IssueData[] = [
-      {
-        number: 1,
-        title: "DISABLED test_a (__main__.suite_a)",
-        html_url: "https://api.github.com/repos/pytorch/pytorch/issues/1",
-        state: "open" as "open" | "closed",
-        body: "random",
-        updated_at: dayjs()
-          .subtract(
-            disableFlakyTestBot.NUM_HOURS_NOT_UPDATED_BEFORE_CLOSING - 1,
-            "hour"
-          )
-          .toString(),
-        author_association: "MEMBER",
-        labels: [],
-      },
-    ];
-
-    await disableFlakyTestBot.handleFlakyTest(flakyTestA, issues, octokit);
-    // Close the disabled issue if the test is not flaky anymore
-    await disableFlakyTestBot.handleNonFlakyTest(
-      nonFlakyTestA,
-      issues,
-      octokit
-    );
-
-    handleScope(scope);
-  });
-});
+import * as flakyBot from "pages/api/flaky-tests/disable";
+import { deepCopy, handleScope } from "test/common";
+import {
+  flakyTestA,
+  flakyTestAcrossJobA,
+  flakyTestB,
+  flakyTestE,
+  mockGetRawTestFile,
+  nonFlakyTestA,
+  nonFlakyTestZ,
+} from "test/flakyBotTests/flakyBotTestsUtils";
+import * as utils from "test/utils";
 
 describe("Disable Flaky Test Bot Unit Tests", () => {
   const octokit = utils.testOctokit();
@@ -553,7 +94,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       },
       flakyTestE,
     ];
-    expect(disableFlakyTestBot.filterOutPRFlakyTests(flakyTests)).toEqual(
+    expect(flakyBot.filterOutPRFlakyTests(flakyTests)).toEqual(
       expectedFlakyTestsOnTrunk
     );
   });
@@ -564,7 +105,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       `# Owner(s): ["module: fft"]\nimport blah;\nrest of file`
     );
     const { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(flakyTestA);
+      await flakyBotUtils.getTestOwnerLabels(flakyTestA);
     expect(additionalErrMessage).toEqual(undefined);
     expect(labels).toEqual(["module: fft", "module: windows", "triaged"]);
 
@@ -577,7 +118,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       `# Owner(s): ["oncall: distributed"]\nimport blah;\nrest of file`
     );
 
-    const { labels } = await disableFlakyTestBot.getTestOwnerLabels(flakyTestA);
+    const { labels } = await flakyBotUtils.getTestOwnerLabels(flakyTestA);
     expect(labels).toEqual([
       "oncall: distributed",
       "module: windows",
@@ -594,7 +135,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     );
 
     const { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(flakyTestA);
+      await flakyBotUtils.getTestOwnerLabels(flakyTestA);
     expect(labels).toEqual(["module: unknown", "module: windows", "triaged"]);
     expect(additionalErrMessage).toEqual(undefined);
 
@@ -608,7 +149,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     );
 
     const { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(flakyTestA);
+      await flakyBotUtils.getTestOwnerLabels(flakyTestA);
     expect(labels).toEqual(["module: windows", "triaged"]);
     expect(additionalErrMessage).toEqual(undefined);
 
@@ -635,7 +176,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       .reply(404);
 
     const { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(flakyTestA);
+      await flakyBotUtils.getTestOwnerLabels(flakyTestA);
     expect(labels).toEqual(["module: windows", "triaged"]);
     expect(additionalErrMessage).toEqual(
       "Error: Error retrieving file_a.py: 404, file_a: 404"
@@ -654,7 +195,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
         Buffer.from(`# Owner(s): ["module: fft"]\nimport blah;\nrest of file`)
       );
     const { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(flakyTestA);
+      await flakyBotUtils.getTestOwnerLabels(flakyTestA);
     expect(labels).toEqual(["module: fft", "module: windows", "triaged"]);
     expect(additionalErrMessage).toEqual(undefined);
 
@@ -671,13 +212,13 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       .reply(404)
       .get(`/pytorch/pytorch/main/test/${flakyTestAcrossJobA.file}`)
       .reply(404)
-      .get(`/pytorch/pytorch/main/test/nn/test_convolution.py`)
+      .get(`/pytorch/pytorch/main/test/${flakyTestAcrossJobA.invoking_file}.py`)
       .reply(
         200,
         Buffer.from(`# Owner(s): ["module: fft"]\nimport blah;\nrest of file`)
       );
     const { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(flakyTestAcrossJobA);
+      await flakyBotUtils.getTestOwnerLabels(flakyTestAcrossJobA);
     expect(labels).toEqual(["module: fft", "module: rocm", "triaged"]);
     expect(additionalErrMessage).toEqual(undefined);
 
@@ -694,7 +235,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     );
 
     let { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(test);
+      await flakyBotUtils.getTestOwnerLabels(test);
     expect(additionalErrMessage).toEqual(undefined);
     expect(labels).toEqual(["module: fft", "oncall: pt2"]);
 
@@ -707,7 +248,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
 
     let scope = mockGetRawTestFile(test.file, `import blah;\nrest of file`);
     let { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(test);
+      await flakyBotUtils.getTestOwnerLabels(test);
     expect(additionalErrMessage).toEqual(undefined);
     expect(labels).toEqual(["oncall: pt2"]);
 
@@ -720,7 +261,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
 
     let scope = mockGetRawTestFile(test.file, `import blah;\nrest of file`);
     let { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(test);
+      await flakyBotUtils.getTestOwnerLabels(test);
     expect(additionalErrMessage).toEqual(undefined);
     expect(labels).toEqual(["module: rocm", "triaged"]);
 
@@ -734,7 +275,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     let scope = mockGetRawTestFile(test.file, `import blah;\nrest of file`);
 
     let { labels, additionalErrMessage } =
-      await disableFlakyTestBot.getTestOwnerLabels(test);
+      await flakyBotUtils.getTestOwnerLabels(test);
     expect(additionalErrMessage).toEqual(undefined);
     expect(labels).toEqual(["module: unknown"]);
 
@@ -742,20 +283,20 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
   });
 
   test("getLatestTrunkJobURL: should return URL of last trunk job if it exists", async () => {
-    expect(disableFlakyTestBot.getLatestTrunkJobURL(flakyTestE)).toEqual(
+    expect(flakyBotUtils.getLatestTrunkJobURL(flakyTestE)).toEqual(
       "https://github.com/pytorch/pytorch/runs/56789876"
     );
   });
 
   test("getLatestTrunkJobURL: should return URL of last job if no trunk instance exists", async () => {
-    expect(disableFlakyTestBot.getLatestTrunkJobURL(flakyTestA)).toEqual(
+    expect(flakyBotUtils.getLatestTrunkJobURL(flakyTestA)).toEqual(
       "https://github.com/pytorch/pytorch/runs/56789876"
     );
   });
 
   test("getIssueTitle: test suite in subclass should not have __main__", async () => {
     expect(
-      disableFlakyTestBot.getIssueTitle(
+      singleDisableIssue.getIssueTitle(
         "test_cool_op_cpu",
         "jit.async.SpecialSuite"
       )
@@ -764,12 +305,12 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
 
   test("getIssueTitle: test suite not in subclass should be prefixed with __main__", async () => {
     expect(
-      disableFlakyTestBot.getIssueTitle("test_cool_op_cpu", "TestLinAlgCPU")
+      singleDisableIssue.getIssueTitle("test_cool_op_cpu", "TestLinAlgCPU")
     ).toEqual("DISABLED test_cool_op_cpu (__main__.TestLinAlgCPU)");
   });
 
   test("getWorkflowJobNames: should zip the workflow and job names of a test", async () => {
-    expect(disableFlakyTestBot.getWorkflowJobNames(flakyTestA)).toEqual([
+    expect(flakyBotUtils.getWorkflowJobNames(flakyTestA)).toEqual([
       "trunk / win-cpu-vs-2019 / test",
       "periodic / win-cuda11.3-vs-2019 / test",
       "periodic / win-cuda11.3-vs-2019 / test",
@@ -781,14 +322,14 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       "periodic / linux-cuda11.1-py3 / test",
       "pull / whatever-whatever-linux / build",
     ];
-    expect(disableFlakyTestBot.getPlatformsAffected(workflowJobNames)).toEqual([
+    expect(flakyBotUtils.getPlatformsAffected(workflowJobNames)).toEqual([
       "linux",
     ]);
   });
 
   test("getPlatformsAffected: should correctly triage workflows of various platforms", async () => {
-    const workflowJobs = disableFlakyTestBot.getWorkflowJobNames(flakyTestE);
-    expect(disableFlakyTestBot.getPlatformsAffected(workflowJobs)).toEqual([
+    const workflowJobs = flakyBotUtils.getWorkflowJobNames(flakyTestE);
+    expect(flakyBotUtils.getPlatformsAffected(workflowJobs)).toEqual([
       "linux",
       "mac",
       "macos",
@@ -798,14 +339,12 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
 
   test("getPlatformsAffected: should correctly triage rocm without linux", async () => {
     const workflowJobs = ["pull / whatever-rocm-linux / build"];
-    expect(disableFlakyTestBot.getPlatformsAffected(workflowJobs)).toEqual([
-      "rocm",
-    ]);
+    expect(flakyBotUtils.getPlatformsAffected(workflowJobs)).toEqual(["rocm"]);
   });
 
   test("getPlatformsAffected: should correctly triage dyanmo and inductor", async () => {
     function expectJobsToDisablePlatforms(jobs: string[], platforms: string[]) {
-      expect(disableFlakyTestBot.getPlatformsAffected(jobs)).toEqual(platforms);
+      expect(flakyBotUtils.getPlatformsAffected(jobs)).toEqual(platforms);
     }
 
     expectJobsToDisablePlatforms(
@@ -848,19 +387,19 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
   });
 
   test("getIssueBodyForFlakyTest: should contain Platforms line", async () => {
-    expect(disableFlakyTestBot.getIssueBodyForFlakyTest(flakyTestA)).toContain(
+    expect(singleDisableIssue.getIssueBodyForFlakyTest(flakyTestA)).toContain(
       "Platforms: "
     );
   });
 
   test("getIssueBodyForFlakyTest: should contain correct examples URL", async () => {
-    expect(disableFlakyTestBot.getIssueBodyForFlakyTest(flakyTestA)).toContain(
+    expect(singleDisableIssue.getIssueBodyForFlakyTest(flakyTestA)).toContain(
       "https://hud.pytorch.org/flakytest?name=test_a&suite=suite_a&limit=100"
     );
   });
 
   test("getIssueBodyForFlakyTest: should contain file info", async () => {
-    expect(disableFlakyTestBot.getIssueBodyForFlakyTest(flakyTestA)).toContain(
+    expect(singleDisableIssue.getIssueBodyForFlakyTest(flakyTestA)).toContain(
       "Test file path: `file_a.py`"
     );
   });
@@ -871,10 +410,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     const flakyTests = [flakyTestA, flakyTestB];
 
     expect(
-      disableFlakyTestBot.filterOutNonFlakyTests(
-        disabledNonFlakyTests,
-        flakyTests
-      )
+      flakyBot.filterOutNonFlakyTests(disabledNonFlakyTests, flakyTests)
     ).toEqual([nonFlakyTestZ]);
   });
 
@@ -902,9 +438,9 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       for (const issue of closed) {
         scope.patch(`/repos/pytorch/pytorch/issues/${issue.number}`).reply(200);
       }
-      expect(
-        await disableFlakyTestBot.dedupFlakyTestIssues(octokit, input)
-      ).toEqual([expected]);
+      expect(await flakyBot.dedupFlakyTestIssues(octokit, input)).toEqual([
+        expected,
+      ]);
       scope.done();
     }
 
@@ -922,7 +458,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
 
   test("parseBody: bodyWithoutPlatforms preserves new lines", async () => {
     expect(
-      parseBody(
+      singleDisableIssue.parseBody(
         "Platforms: win, rocm\r\nHello\r\n\nMy name is\n\n```a;dlskfja\ndklfj```"
       )
     ).toEqual({
@@ -931,7 +467,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
       bodyWithoutPlatforms:
         "\r\nHello\r\n\nMy name is\n\n```a;dlskfja\ndklfj```",
     });
-    expect(parseBody("Platforms: win, rocm\r\n")).toEqual({
+    expect(singleDisableIssue.parseBody("Platforms: win, rocm\r\n")).toEqual({
       platformsToSkip: ["rocm", "win"],
       invalidPlatforms: [],
       bodyWithoutPlatforms: "\r\n",
@@ -965,7 +501,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
         )
         .reply(200, {});
 
-      await disableFlakyTestBot.updateExistingIssueForFlakyTest(
+      await singleDisableIssue.updateExistingIssueForFlakyTest(
         octokit,
         defaultIssue,
         test
@@ -1001,7 +537,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
         })
         .reply(200, {});
 
-      await disableFlakyTestBot.updateExistingIssueForFlakyTest(
+      await singleDisableIssue.updateExistingIssueForFlakyTest(
         octokit,
         issue,
         test
@@ -1041,7 +577,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
         })
         .reply(200, {});
 
-      await disableFlakyTestBot.updateExistingIssueForFlakyTest(
+      await singleDisableIssue.updateExistingIssueForFlakyTest(
         octokit,
         issue,
         test
@@ -1083,7 +619,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
         })
         .reply(200, {});
 
-      await disableFlakyTestBot.updateExistingIssueForFlakyTest(
+      await singleDisableIssue.updateExistingIssueForFlakyTest(
         octokit,
         issue,
         test
@@ -1127,7 +663,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
         })
         .reply(200, {});
 
-      await disableFlakyTestBot.updateExistingIssueForFlakyTest(
+      await singleDisableIssue.updateExistingIssueForFlakyTest(
         octokit,
         issue,
         test
@@ -1167,7 +703,7 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
         })
         .reply(200, {});
 
-      await disableFlakyTestBot.updateExistingIssueForFlakyTest(
+      await singleDisableIssue.updateExistingIssueForFlakyTest(
         octokit,
         issue,
         test
